requirements.txt
# Connection to CAN
pyserial==3.3

# Figley
pyfiglet

# Sensors
hokuyolx

# Mapping
matplotlib
numpy